+++ Player.h
@@ -1,7 +1,5 @@
 class Player {
 public:
     bool tp = false;
-    int x,y;
-
-
+    float x,y;
 };
